aircraft_manager/pyproject.toml
[tool.poetry]
name = "aircraft-manager"
version = "0.1.0"
description = ""
authors = ["Greg <100309099+Greg75@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
psycopg2 = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-env = "^1.1.5"
pytest-mock = "^3.14.0"
pytest-timeout = "^2.3.1"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
